+++ src/main/java/com/example/AdSponsor/model/Product.java
@@ -2,19 +2,21 @@ package com.example.AdSponsor.model;
 import org.springframework.data.relational.core.mapping.Table;
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
+
 @Table(name = "PRODUCTS")
 public class Product {
     @Id
     private Integer id;
     private String title;
     private String category;
-    private double price;
+    private BigDecimal price;
 
     public Product() {
         // Default constructor
     }
 
-    public Product(String title, String category, double price) {
+    public Product(String title, String category, BigDecimal price) {
         this.title = title;
         this.category = category;
         this.price = price;
@@ -44,11 +46,11 @@ public class Product {
         this.category = category;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
